% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\name{render_site}
\alias{render_site}
\title{Render Site}
\usage{
render_site(..., output_dir = "_site")
}
\arguments{
\item{...}{additional arguments to be passed to bookdown::render_book .}

\item{output_dir}{string indicating directory where to render html files for website. Default is "_site".}
}
\description{
Function that would create website using bookdown::render_book.
}
