% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_path}
\alias{report_path}
\title{Construct Report Path}
\usage{
report_path(report)
}
\arguments{
\item{report}{character string indicating name of the report without extension.}
}
\value{
caracther string indicating the path to the given report.

None
}
\description{
Function to construct a path for a given report
}
